Update AMAmigrationTracker.json
The query that we have currently in line 367 is incorrect. The join is happening only for the AzureVMs which results in all HybridCompute machines to be shown in the results regardless if they have one, or two agents. I am not sure why the join is done but in my point of view it is completely unnecessary. 

The following query gives the correct results:
resources
| where type has 'microsoft.compute/virtualmachines/extensions' or  type has 'microsoft.hybridcompute/machines/extensions'
| where name has 'MicrosoftMonitoringAgent' or name has 'OmsAgentForLinux' or name has 'AzureMonitorWindowsAgent' or name has 'AzureMonitorLinuxAgent'
| extend AzureVM = extract('virtualMachines/(.*)/extensions',1,id), ArcVM = extract('machines/(.*)/extensions',1,id)
| summarize count() by AzureVM=tolower(AzureVM), ArcVM=tolower(ArcVM), subscriptionId, resourceGroup 
| extend hasBoth = iff(count_ > 1, 'Yes', 'No')
| where hasBoth == "Yes"
| summarize count()